backend/requirements.txt
APScheduler==3.9.1
Flask==2.3.2
Flask-CORS==4.0.0
Flask-Dance==7.1.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==2.8.1
Werkzeug==2.2.2
celery==5.3.1
chatterbot==1.0.8
cryptography==39.0.1
facebook-sdk==4.0.0
fastapi==0.95.2
flask-swagger-ui==4.11.1
gTTS==2.3.2
google-api-python-client==2.86.0
marshmallow==3.19.0
matplotlib==3.7.2
minio==7.1.15
moviepy==1.0.3
pandas==2.0.3
pillow==9.5.0
psycopg2-binary==2.9.5
pyotp==2.8.0
python-dotenv==1.0.0
pytesseract==0.3.10
redis==5.0.0
scikit-learn==1.2.2
sqlalchemy==1.4.41
SpeechRecognition==3.10.0
spotipy==4.8.0
tensorflow==2.12.0
transformers==4.26.1
tweepy==4.14.0
twilio==8.3.0
uvicorn==0.22.0
wikipedia-api==0.6.0